requirements.txt
asgiref==3.3.4
Django==3.2.3
django-view-table==1.0.1
djangorestframework==3.12.4
psycopg2-binary==2.8.6
python-decouple==3.4
pytz==2021.1
sqlparse==0.4.1
